Fix some circular imports.
