+++ splitgraph/ingestion/singer/__init__.py
@@ -3,7 +3,7 @@
 import click
 
 from splitgraph.commandline.common import ImageType
-from splitgraph.ingestion.singer._utils import prepare_new_image
+from splitgraph.hooks.data_source.base import prepare_new_image
 
 
 @click.group(name="singer")
